conveyor_backend/requirements.txt
django
djangorestframework
django-cors-headers
opencv-python-headless
numpy
Pillow
dj-database-url
psycopg2-binary
channels
channels-redis
daphne
ultralytics==8.2.77
opencv-python
django-extensions